Move distance compute functions and ```struct Coords``` into separate files.
These functions compute distances using ```BusManager``` data.
