+++ src/bus_manager.cpp
@@ -1,7 +1,7 @@
 #include "bus_manager.h"
+#include "distance_computer.h"
 
 #include <algorithm>
-#include <cmath>
 #include <optional>
 #include <set>
 #include <string>
@@ -13,25 +13,6 @@
 #include <vector>
 
 namespace {
-// Used the solution from https://stackoverflow.com/a/70429229.
-// Uses Haversine formula https://en.wikipedia.org/wiki/Haversine_formula
-// with asin expressed with atan2 for better accuracy.
-double CalculateDistance(rm::Coords lhs, rm::Coords rhs) {
-  constexpr double earth_radius_m = 6371.0 * 1000.0;
-
-  const double d_lat = rhs.latitude - lhs.latitude;
-  const double d_long = rhs.longitude - lhs.longitude;
-
-  double a =
-      std::sin(d_lat / 2) * std::sin(d_lat / 2) + std::sin(d_long / 2) *
-          std::sin(d_long / 2) *
-          std::cos(lhs.latitude) *
-          std::cos(rhs.latitude);
-  double c = earth_radius_m * 2 * std::atan2(std::sqrt(a), std::sqrt(1 - a));
-
-  return c;
-}
-
 int ComputeUniqueCount(std::vector<std::string> stops) {
   std::sort(begin(stops), end(stops));
   return std::distance(stops.begin(),
@@ -80,8 +61,8 @@ BusManager::BusManager(std::vector<PostRequest> requests) {
   }
 
   for (auto &[bus, bus_info] : bus_info_) {
-    double geo_dist = ComputeGeoDistance(bus_info.stops);
-    bus_info.distance = ComputeRoadDistance(bus_info.stops);
+    double geo_dist = ComputeGeoDistance(bus_info.stops, stop_info_);
+    bus_info.distance = ComputeRoadDistance(bus_info.stops, stop_info_);
     bus_info.unique_stop_count = ComputeUniqueCount(bus_info.stops);
     bus_info.curvature = bus_info.distance / geo_dist;
   }
@@ -93,7 +74,7 @@ BusManager::BusManager(std::vector<PostRequest> requests) {
   }
 }
 
-void BusManager::AddStop(const std::string &stop, Coords coords,
+void BusManager::AddStop(const std::string &stop, sphere::Coords coords,
                          const std::map<std::string, int> &stops) {
   constexpr double k = 3.1415926535 / 180;
   auto &info = stop_info_[stop];
@@ -136,28 +117,5 @@ std::optional<StopResponse> BusManager::GetStopInfo(const std::string &stop) con
     return std::nullopt;
 
   return StopResponse{it->second.buses};
-}
-
-double BusManager::ComputeGeoDistance(const std::vector<std::string> &stops) const {
-  double distance = 0;
-  for (int i = 1; i < stops.size(); ++i) {
-    distance += CalculateDistance(stop_info_.at(stops[i - 1]).coords,
-                                  stop_info_.at(stops[i]).coords);
-  }
-  return distance;
-}
-
-double BusManager::ComputeRoadDistance(const std::vector<std::string> &stops) const {
-  double distance = 0;
-  for (int i = 1; i < stops.size(); ++i) {
-    auto &dists = stop_info_.at(stops[i - 1]).dists;
-
-    if (auto found = dists.find(stops[i]); found != dists.end())
-      distance += found->second;
-    else
-      distance += CalculateDistance(stop_info_.at(stops[i - 1]).coords,
-                                    stop_info_.at(stops[i]).coords);
-  }
-  return distance;
 }
 }
